// src/app/models/UserModel.ts
export class UserModel {
    // email: string;
    // emailConfirmed: boolean;
    // password: string;
    // phoneNumber: string;
    // firstName: string;
    // lastName : string;
    // county : string;
    // city : string;
    // zipcode: string;
    firstName: string;
    lastName: string;
    email: string;
    country: string;
    city: string;
    zipCode: string;
    password: string;
    
}